pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "restaurant-rec"
version = "0.1.0"
description = "Zomato-style restaurant recommendation (phase 1 catalog ingest)"
requires-python = ">=3.9"
dependencies = [
  "datasets>=2.14.0",
  "fastapi>=0.115.0",
  "google-genai>=1.0.0",
  "pandas>=2.0.0",
  "pyarrow>=14.0.0",
  "pydantic>=2.0.0",
  "python-dotenv>=1.0.0",
  "pyyaml>=6.0",
  "streamlit>=1.30.0",
  "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = ["httpx>=0.27.0", "pytest>=7.0.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
